+++ .vscode/launch.json
@@ -8,7 +8,8 @@
       "name": "Python Debugger: Current File",
       "type": "debugpy",
       "request": "launch",
-      "program": "__main__.py",
+      "program": "examples/example_mid_level.py",
+      "args": ["COM3"],
       "console": "integratedTerminal"
     }
   ]
