+++ README.md
@@ -15,3 +15,9 @@ Coming soon
 ## Dependencies
 
 PySerial (https://pypi.org/project/pyserial/) (py -m pip install pyserial)
+
+## Build library
+
+python -m pip install --upgrade build
+python -m build
+pip install .\dist\science_mode_4-0.0.0-py3-none-any.whl --force-reinstall
