+++ src/science_mode_4/device.py
@@ -4,7 +4,7 @@ from enum import Enum
 from typing import Type
 from .low_level.low_level_layer import LayerLowLevel
 from .mid_level.mid_level_layer import LayerMidLevel
-from .protocol.types import StimStatus
+from .types.stim_status import StimStatus
 from .layer import Layer
 from .general.general_layer import LayerGeneral
 from .protocol.packet_factory import PacketFactory
