+++ src/science_mode_4/layer.py
@@ -1,6 +1,6 @@
 """Provides base class for all ScienceMode layers"""
 
-from .protocol.types import ResultAndError
+from .types.result_and_error import ResultAndError
 from .protocol.packet_factory import PacketFactory
 from .protocol.packet_number_generator import PacketNumberGenerator
 from .utils.connection import Connection
