+++ src/science_mode_4/mid_level/mid_level_init.py
@@ -2,7 +2,7 @@
 
 from ..protocol.commands import Commands
 from ..protocol.packet import Packet, PacketAck
-from ..protocol.types import ResultAndError
+from ..types.result_and_error import ResultAndError
 from ..utils.byte_builder import ByteBuilder
 
 class PacketMidLevelInit(Packet):
